requirements.txt
dnspython>=2.4.0
python-whois>=0.9.0
requests>=2.31.0
rich>=13.0.0
